WIP: editable list
