+++ index.html
@@ -17,17 +17,6 @@
 
     <section>
       <h2>Distractions</h2>
-
-      <form class="add-distraction">
-        <input type="text" />
-        <button type="submit">Add</button>
-      </form>
-
-      <ul class="distractions">
-        <template class="distraction-template">
-          <li><span></span><button>&times;</button></li>
-        </template>
-      </ul>
     </section>
   </main>
 </body>
