+++ lib/countdown-timer/main.js
@@ -3,18 +3,12 @@
 import {getTemplate} from '../jt-component/main.js'
 
 export class CountdownTimer extends HTMLElement {
-  constructor () {
-    super()
-    this._intervalReference = null
-
-    this._templatePromise = getTemplate('./lib/countdown-timer/index.html')
-  }
-
   // Specify observed attributes so that `attributeChangedCallback` will work
   static get observedAttributes () { return ['duration', 'start-text', 'stop-text'] }
 
   async connectedCallback () {
-    const template = await this._templatePromise
+    this._intervalReference = null
+    const template = await getTemplate('./lib/countdown-timer/index.html')
 
     this._timeDisplayElement = template.element.getElementById('time-display')
