+++ lib/editable-list/index.html
@@ -1,0 +1,21 @@
+<!DOCTYPE html>
+<html>
+<head>
+  <title>Editable List</title>
+</head>
+<body>
+  <template id="form-and-list">
+    <form>
+      <input type="text" />
+      <!-- TODO: add custom attribute to change this button -->
+      <button type="submit">Add</button>
+    </form>
+
+    <ul></ul>
+  </template>
+
+  <template id="list-item">
+    <li><span></span><button>&times;</button></li>
+  </template>
+</body>
+</html>
