+++ lib/jt-component/main.js
@@ -7,7 +7,24 @@ export async function getTemplate (templateUrl) {
 
   const templateDocument = (new DOMParser())
     .parseFromString(htmlText, 'text/html')
-  const templateElement = templateDocument.querySelector('template')
+
+  // TODO: put the below items in a Map
+  // How it works:
+  // - Add index->element to Map
+  // - If the item has an ID then add id->element to Map
+  // In this way people can access the element by ID or index
+  // This may be bad for iteration, but we can cross that bridge when we get
+  // there.
+  const templateElements = new Map()
+  for (const element of templateDocument.querySelectorAll('template')) {
+    
+
+    if (element) {
+
+    }
+  }
+
+  const templateElement = templateDocument.querySelectorAll('template')
     .content
     .cloneNode(true)
   const templateStylesheet = templateDocument
