+++ script.js
@@ -8,8 +8,6 @@ const COUNTDOWN_SECONDS = 25 * 60 // 25 minutes
 const notificationSound = new Audio()
 
 const countdown = document.getElementById('countdown')
-const distractionForm = document.querySelector('.add-distraction')
-const distractionsList = document.querySelector('.distractions')
 
 init()
 
@@ -21,34 +19,10 @@ function init () {
 
   customElements.define('countdown-timer', CountdownTimer)
 
-  distractionForm.addEventListener('submit', function (event) {
-    event.preventDefault()
-
-    const inputElement = distractionForm.querySelector('input')
-    addDistraction(inputElement.value)
-
-    distractionForm.reset()
-    inputElement.focus()
-  })
-
   if (Notification.permission === 'default') {
     // Request notification permission if it hasn't been explicitly granted or denied.
     Notification.requestPermission()
   }
-}
-
-function addDistraction (distraction) {
-  // TODO: add an auto-complete feature based on what you previously typed in here.
-  const liElement = distractionsList
-    .querySelector('.distraction-template')
-    .content
-    .firstElementChild
-    .cloneNode(true)
-
-  liElement.querySelector('span').appendChild(document.createTextNode(distraction))
-  liElement.querySelector('button').addEventListener('click', () => liElement.remove())
-
-  distractionsList.appendChild(liElement)
 }
 
 function showNotification () {
